requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
drf-generators==0.5.0
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.2.0 
pytz==2021.1
sqlparse==0.4.1
waitress==2.0.0
whitenoise==5.2.0 
